app/requirements.txt
Django==4.1.2
djangorestframework==3.14.0
django-cors-headers==3.13.0